Optimize dependencies
For now docker-compose fails building the web image because installing and building the libsass
python package takes too much memory. I created an image with that package already installed so
that the dependency would be marked as installed, speeding up the build slightly and allowing
it to be constructed in higher-usage environments.