Fix settings environment lookup
If a setting was looked up by a string name, and the name
didn't occur in the environment, then it would loop through
each character of the string and try that in the environment.

This doesn't work, because _ is generally the name of the currently
running binary, which will return a value, and thus implode.

Fix this by not looping through if the lookup key is a string.